+++ systemtest/src/test/java/io/strimzi/systemtest/SecurityST.java
@@ -37,7 +37,6 @@ import io.strimzi.systemtest.utils.kubeUtils.objects.PodUtils;
 import io.strimzi.systemtest.utils.kubeUtils.objects.SecretUtils;
 import io.strimzi.systemtest.utils.specific.MetricsUtils;
 import io.strimzi.test.TestUtils;
-import io.strimzi.test.executor.Exec;
 import io.strimzi.test.k8s.exceptions.KubeClusterException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -46,22 +45,17 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.Base64;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -1302,65 +1296,6 @@ class SecurityST extends BaseST {
         assertThat(sent, is(messagesCount));
         received = internalKafkaClient.receiveMessagesTls(topicName, NAMESPACE, CLUSTER_NAME, userName, messagesCount, "TLS", CONSUMER_GROUP_NAME + "-" + rng.nextInt(Integer.MAX_VALUE));
         assertThat(received, is(sent));
-    }
-
-    @Test
-    void testLoadBalancerSourceRanges() throws IOException, InterruptedException, ExecutionException, TimeoutException {
-
-        String networkInterfaces = Exec.exec("ip", "route").out();
-        Pattern ipv4InterfacesPattern = Pattern.compile("[0-9]+.[0-9]+.[0-9]+.[0-9]+\\/[0-9]+ dev (eth0|enp11s0u1).*");
-        Matcher ipv4InterfacesMatcher = ipv4InterfacesPattern.matcher(networkInterfaces);
-
-        ipv4InterfacesMatcher.find();
-        LOGGER.info(ipv4InterfacesMatcher.group(0));
-        String correctNetworkInterface = ipv4InterfacesMatcher.group(0);
-
-        String[] correctNetworkInterfaceStrings = correctNetworkInterface.split(" ");
-
-        String ipWithPrefix = correctNetworkInterfaceStrings[0];
-
-        LOGGER.info("Network address of machine with associated prefix is {}", ipWithPrefix);
-
-        KafkaResource.kafkaPersistent(CLUSTER_NAME, 3)
-            .editSpec()
-                .editKafka()
-                    .editListeners()
-                        .withNewKafkaListenerExternalLoadBalancer()
-                            .withTls(false)
-                        .endKafkaListenerExternalLoadBalancer()
-                    .endListeners()
-                    .withNewTemplate()
-                        .withNewExternalBootstrapService()
-                                .withLoadBalancerSourceRanges(Collections.singletonList(ipWithPrefix))
-                        .endExternalBootstrapService()
-                        .withNewPerPodService()
-                            .withLoadBalancerSourceRanges(ipWithPrefix)
-                        .endPerPodService()
-                    .endTemplate()
-                .endKafka()
-            .endSpec()
-            .done();
-
-        Future<Integer> producer = externalBasicKafkaClient.sendMessages(TOPIC_NAME, NAMESPACE, CLUSTER_NAME, MESSAGE_COUNT);
-        Future<Integer> consumer = externalBasicKafkaClient.receiveMessages(TOPIC_NAME, NAMESPACE, CLUSTER_NAME, MESSAGE_COUNT);
-
-        assertThat(producer.get(Constants.GLOBAL_CLIENTS_TIMEOUT, TimeUnit.MILLISECONDS), is(MESSAGE_COUNT));
-        assertThat(consumer.get(Constants.GLOBAL_CLIENTS_TIMEOUT, TimeUnit.MILLISECONDS), is(MESSAGE_COUNT));
-
-        String invalidNetworkAddress = "255.255.255.111/30";
-
-        LOGGER.info("Replacing Kafka CR invalid load-balancer source range to {}", invalidNetworkAddress);
-
-        KafkaResource.replaceKafkaResource(CLUSTER_NAME, kafka ->
-            kafka.getSpec().getKafka().getTemplate().getExternalBootstrapService().setLoadBalancerSourceRanges(Collections.singletonList(invalidNetworkAddress))
-        );
-
-        LOGGER.info("Expecting that clients will not be able to connect to external load-balancer service cause of invalid load-balancer source range.");
-
-        assertThrows(ExecutionException.class, () -> {
-            Future<Integer> invalidProducer = externalBasicKafkaClient.sendMessages(TOPIC_NAME, NAMESPACE, CLUSTER_NAME, MESSAGE_COUNT);
-            Future<Integer> invalidConsumer = externalBasicKafkaClient.receiveMessages(TOPIC_NAME, NAMESPACE, CLUSTER_NAME, MESSAGE_COUNT * 2);
-        });
     }
 
     @BeforeAll
